requirements.txt
PyMySQL==0.7.9
paho-mqtt==1.6.1
requests==2.27.1
